+++ Content.Client/Content.Client.csproj
@@ -26,7 +26,7 @@
   <ItemGroup>
     <Folder Include="Clothing\Components\" />
     <Folder Include="Spawners\" />
-    <Folder Include="_White\VisionLimit\Overlays\System\" />
+    <Folder Include="_White\VisionLimiter\Overlays\System\" />
   </ItemGroup>
   <ItemGroup>
     <Compile Update="Lobby\UI\LobbyCharacterPanel.xaml.cs">
