--- Content.Shared/_White/VisionLimit/Components/VisionLimitComponent.cs
@@ -1,8 +1,0 @@
-﻿namespace Content.Shared._White.VisionLimit.Components;
-
-[RegisterComponent]
-public sealed partial class VisionLimitComponent: Component
-{
-    [DataField]
-    public float Radius; // in tiles, 0 = no limit
-}
